Spi: new methods for time critical applications
Most of the time there is only one device on the bus. Sometimes overhead
of obtaining the mutex before every transmission is critical. These new
methods allow to take semaphore for extended periods of time, e.g. to
cover the whole communication session. They have to be used with care if
there are multipple devices on the bus. Even if not, it is preferred to
enclose communication sections with obtainMutex() and giveMutex(), because
you never know when and you will add a second device to the bus.

Added convenience methods:
readByte(*data), writeByte(data), shiftByte(data, *data)

	modified:   drivers/Spi.cpp
	modified:   drivers/Spi.h